backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20
langchain_community==0.2.19
nltk==3.9.1
pydantic_settings==2.8.1
python-terrier==0.13.0
requests==2.32.3
beanie==1.29.0
